fix: remove functional debt and align fresh-install schema (#214)

+++ lib/screens/product_detail_screen.dart
@@ -692,7 +692,7 @@ class _ProductDetailScreenState extends ConsumerState<ProductDetailScreen> {
           );
           if (system == UnitSystem.imperial) {
             final converted = UnitConverter.displayUnit(
-              medium.toDouble(),
+              medium,
               'g',
               UnitSystem.imperial,
               weightPref: settings.preferredWeightUnit,
